+++ src/MinFraud.php
@@ -297,13 +297,19 @@ class MinFraud extends MinFraud\ServiceClient implements \JsonSerializable
      *                                            are:
      *                                            - `account_creation`
      *                                            - `account_login`
+     *                                            - `credit_application`
      *                                            - `email_change`
+     *                                            - `fund_transfer`
      *                                            - `password_reset`
      *                                            - `payout_change`
      *                                            - `purchase`
      *                                            - `recurring_purchase`
      *                                            - `referral`
      *                                            - `survey`
+     * @param string|null          $party         The party submitting the transaction. The valid values
+     *                                            are:
+     *                                            - `agent`
+     *                                            - `customer`
      *
      * @return MinFraud A new immutable MinFraud object. This object is a clone of
      *                  the original with additional data.
@@ -317,6 +323,7 @@ class MinFraud extends MinFraud\ServiceClient implements \JsonSerializable
         ?string $time = null,
         ?string $transactionId = null,
         ?string $type = null,
+        ?string $party = null,
     ): self {
         if (\count($values) !== 0) {
             if (\func_num_args() !== 1) {
@@ -324,12 +331,20 @@ class MinFraud extends MinFraud\ServiceClient implements \JsonSerializable
                     'You may only provide the $values array or named arguments, not both.',
                 );
             }
+            $party = $this->remove($values, 'party');
             $shopId = $this->remove($values, 'shop_id');
             $time = $this->remove($values, 'time');
             $transactionId = $this->remove($values, 'transaction_id');
             $type = $this->remove($values, 'type');
 
             $this->verifyEmpty($values);
+        }
+
+        if ($party !== null) {
+            if (!\in_array($party, ['agent', 'customer'], true)) {
+                $this->maybeThrowInvalidInputException("$party is not a valid party");
+            }
+            $values['party'] = $party;
         }
 
         if ($shopId !== null) {
@@ -354,7 +369,9 @@ class MinFraud extends MinFraud\ServiceClient implements \JsonSerializable
             if (!\in_array($type, [
                 'account_creation',
                 'account_login',
+                'credit_application',
                 'email_change',
+                'fund_transfer',
                 'password_reset',
                 'payout_change',
                 'purchase',
@@ -753,6 +770,18 @@ class MinFraud extends MinFraud\ServiceClient implements \JsonSerializable
      * @param bool|null            $wasAuthorized The authorization outcome from the payment
      *                                            processor. If the transaction has not yet been
      *                                            approved or denied, do not include this field.
+     * @param string|null          $method        The payment method associated with the transaction.
+     *                                            The valid values are:
+     *                                            - `bank_debit`
+     *                                            - `bank_redirect`
+     *                                            - `bank_transfer`
+     *                                            - `buy_now_pay_later`
+     *                                            - `card`
+     *                                            - `crypto`
+     *                                            - `digital_wallet`
+     *                                            - `gift_card`
+     *                                            - `real_time_payment`
+     *                                            - `rewards`
      *
      * @return MinFraud A new immutable MinFraud object. This object is
      *                  a clone of the original with additional data.
@@ -762,6 +791,7 @@ class MinFraud extends MinFraud\ServiceClient implements \JsonSerializable
         ?string $declineCode = null,
         ?string $processor = null,
         ?bool $wasAuthorized = null,
+        ?string $method = null,
     ): self {
         if (\count($values) !== 0) {
             if (\func_num_args() !== 1) {
@@ -771,6 +801,7 @@ class MinFraud extends MinFraud\ServiceClient implements \JsonSerializable
             }
 
             $declineCode = $this->remove($values, 'decline_code');
+            $method = $this->remove($values, 'method');
             $processor = $this->remove($values, 'processor');
             $wasAuthorized = $this->remove($values, 'was_authorized', ['boolean']);
 
@@ -779,6 +810,24 @@ class MinFraud extends MinFraud\ServiceClient implements \JsonSerializable
 
         if ($declineCode !== null) {
             $values['decline_code'] = $declineCode;
+        }
+
+        if ($method !== null) {
+            if (!\in_array($method, [
+                'bank_debit',
+                'bank_redirect',
+                'bank_transfer',
+                'buy_now_pay_later',
+                'card',
+                'crypto',
+                'digital_wallet',
+                'gift_card',
+                'real_time_payment',
+                'rewards',
+            ], true)) {
+                $this->maybeThrowInvalidInputException("$method is not a valid payment method");
+            }
+            $values['method'] = $method;
         }
 
         if ($processor !== null) {